NEW: drivers/vl6180x: Multi-sensor support
At power-up the VL6180X has a default I2C address, which is the
same for all VL6180X, which means we can't have multiple VL6180X
on the bus by default without collision. But the I2C address is
configurable and there is a pin (XSHUT) to put the sensor in HW
standby. So if we put all the sensors in HW standby and then bring
them up one by one setting a unique address for each one, we can
use multiple of them at the same time. This approach is described
in app note AN4478.

Since we need to toggle GPIO, create a simple GPIO driver to hide
the mcu specifics to keep the sensor layer portable.