ENH: drivers/vl53l0x: Set SPADs from NVM
ST stores a SPAD configuration in non volatile memory (NVM) during production.
We can read this and configure the SPADs accordingly. This should be a good-enough
configuration for the normal use-case where there is no cover glass.

For an even better configuration, we could run the SPAD management/calibration
ourselves, and the API code from ST contains example code on how to do that,
but it requires tedious code, so I don't include it here.